Validation logic moved to HospitalAPI and used custom exceptions
