--- src/HospitalAPI/ErrorHandling/Exceptions/BadRequestException.cs
@@ -1,11 +1,0 @@
-﻿using System.Net;
-
-namespace HospitalAPI.ErrorHandling.Exceptions
-{
-    public class BadRequestException: BaseException
-    {
-        public BadRequestException(string message) : base(message,(int)HttpStatusCode.BadRequest)
-        {
-        }
-    }
-}
