+++ src/HospitalAPI/Startup.cs
@@ -3,6 +3,7 @@ using HospitalLibrary.Appointments.Interfaces;
 using System;
 using HospitalAPI.Emails;
 using HospitalAPI.ErrorHandling;
+using HospitalAPI.Validators;
 using HospitalLibrary.Doctors;
 using HospitalLibrary.Doctors.Interfaces;
 using HospitalLibrary.Feedbacks;
@@ -14,7 +15,6 @@ using HospitalLibrary.Rooms.Interfaces;
 using HospitalLibrary.Settings;
 using HospitalLibrary.Shared.Interfaces;
 using HospitalLibrary.Shared.Repository;
-using HospitalLibrary.Shared.Service;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
