+++ src/HospitalAPI/appsettings.json
@@ -8,7 +8,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "HospitalDb": "User Id=postgres;Password=ftn;Server=localhost;Port=5432;Database=HospitalDb"
+    "HospitalDb": "User Id=postgres;Password=postgresssql;Server=localhost;Port=5432;Database=HospitalDb"
   }
 }
