+++ src/HospitalLibrary/HospitalLibrary.csproj
@@ -31,6 +31,7 @@
     <Folder Include="MedicalRecords\Service\" />
     <Folder Include="MedicalRecords\Repository\" />
     <Folder Include="Migrations" />
+    <Folder Include="Shared\Service" />
   </ItemGroup>
 
 </Project>
